+++ mirrord/operator/src/lib.rs
@@ -1,0 +1,7 @@
+#![feature(once_cell)]
+
+/// Operator Setup functinality
+#[cfg(feature = "setup")]
+pub mod setup;
+
+pub mod license;